backend/requirements.txt
livekit-agents~=1.5
livekit-plugins-openai~=1.5
livekit-plugins-deepgram~=1.5
livekit-plugins-elevenlabs~=1.5
livekit-plugins-silero~=1.5
livekit-plugins-turn-detector~=1.5
livekit-api~=1.1
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-openai
llama-index-llms-openai
llama-index-readers-file
chromadb
fastapi
uvicorn[standard]
python-dotenv
httpx
pymupdf
pyyaml
pytest
pytest-asyncio
